move closest coi
